be/requirements.txt
aniso8601==9.0.1
attrs==21.4.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
Flask==2.1.2
Flask-Login==0.6.1
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.9
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.15.0
marshmallow3-annotations==1.0.0
oauthlib==3.2.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
PyJWT==2.3.0
pymongo==4.1.1
pyparsing==3.0.8
pytest==7.1.2
pytz==2022.1
requests==2.27.1
six==1.16.0
toml==0.10.2
tomli==2.0.1
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
